+++ changelog.md
@@ -1,3 +1,19 @@
+# 1.2.0 (13/01/2026)
+Création des commandes:
+- music_list (info / other) : Liste des musiques sur le stockage du REMI.
+- play_selectmusic (action / liste) : Joue une musique listée 
+- event_list (info / other) : Liste des reveils programmés sur REMI (HH:MM - Nom: 0=disable,1=enable)
+- noise_notification_subscribers (info / numeric) : Nombre d'appareils abonné aux notifications en cas de bruit détecté.
+- event_enable (action / list) : Active un reveil sélectionné dans la liste
+- event_disable (action / list) : Désactive un reveil sélectionné dans la liste
+- set_face (action / list) : Appliquer un visage sélectionné dans la liste
+- gettime_wayback_max_s (info / numeric) : Si remonté par l'API, délai max avant récupération de l'heure
+- call_wayback_max_s (info / numeric) : Si remonté par l'API, délai max avant reconnexion
+- gettime_shift_s (info / numeric) : Si remonté par l'API, décalage de temps
+- day_reconnection_count (info / numeric) : Si remonté par l'API, nombre de reconnexions du jour
+- day_disconnection_time (info / numeric) : Si remonté par l'API, nombre de déconnexions du jour
+- alarm_XXXXX (info / binary) : Il s'agit des events/reveils paramétrés. 0=disable, 1=enable
+
 # 1.1.8 (12/01/2026)
 Correction d'un bug impectant la lecture et arret des musiques.